feature/add-AsEntityUpdateEventListener-and-AsEntityCreatedEventListener-and-AsEntityDeletedEventListener-attributes-to-EasyDoctrine
- Code review fixes
