+++ packages/EasyTest/src/EasyEventDispatcher/Dispatcher/EventDispatcherStub.php
@@ -33,6 +33,8 @@ final class EventDispatcherStub implements EventDispatcherInterface
 
     public function dispatch(object $event, ?string $eventName = null): object
     {
+        $this->events[] = $event;
+        
         $callback = $this->dispatchCallbacks[$event::class] ?? null;
 
         if ($callback !== null) {
